requirements.txt
fastapi>=0.100.0
uvicorn==0.23.2
openai>=1.3.0
python-dotenv>=0.19.0,<0.20.0
pydantic>=2.0.0
python-multipart>=0.0.5,<0.0.6
httpx

